Kingdom #3 - Added message when you enter claim and more
tried my best because i'm dumb and i don't push enough, there's now a message that shows when you enter a land. I also started working on checking if a player is from the town the claim is owned by before breaking the block. It's crashing for now but i'm working on it.